[tests] Move DDP utils for tests
