+++ tests/dummy/utils/ddp.py
@@ -1,5 +1,8 @@
 import functools
+import os
 from typing import Any, Callable
+
+import torch
 
 
 def retry_on_file_not_found(num_retries: int, enable: bool = True) -> Callable:
@@ -31,3 +34,21 @@ def retry_on_file_not_found(num_retries: int, enable: bool = True) -> Callable:
         return _wrapper
 
     return wrapper
+
+
+def set_ddp_environment(
+    rank: int,
+    world_size: int,
+    port: int,
+) -> None:
+    os.environ["LOCAL_RANK"] = str(rank)
+    os.environ["RANK"] = str(rank)
+    os.environ["WORLD_SIZE"] = str(world_size)
+    os.environ["MASTER_ADDR"] = "localhost"
+    os.environ["MASTER_PORT"] = str(port)
+    os.environ["TORCH_CPP_LOG_LEVEL"] = "INFO"
+    os.environ["TORCH_DISTRIBUTED_DEBUG"] = "DETAIL"
+
+    num_threads = torch.get_num_threads()
+    num_threads = max(num_threads // world_size, 1)
+    torch.set_num_threads(num_threads)
